@@ -3,46 +3,53 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "bootstrap": "^5.1.0",
-    "http-proxy-middleware": "^0.19.1",
-    "jquery": "^3.5.1",
+    "bootstrap": "^5.2.0",
+    "http-proxy-middleware": "^2.0.6",
+    "jquery": "^3.6.0",
     "merge": "^2.1.1",
     "oidc-client": "^1.11.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "reactstrap": "^8.9.0",
-    "rimraf": "^2.6.2",
-    "web-vitals": "^0.2.4",                
-    "workbox-background-sync": "^5.1.3",   
-    "workbox-broadcast-update": "^5.1.3",  
-    "workbox-cacheable-response": "^5.1.3",
-    "workbox-core": "^5.1.3",              
-    "workbox-expiration": "^5.1.3",        
-    "workbox-google-analytics": "^5.1.3",  
-    "workbox-navigation-preload": "^5.1.3",
-    "workbox-precaching": "^5.1.3",        
-    "workbox-range-requests": "^5.1.3",    
-    "workbox-routing": "^5.1.3",           
-    "workbox-strategies": "^5.1.3",        
-    "workbox-streams": "^5.1.3"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "reactstrap": "^9.1.3",
+    "rimraf": "^3.0.2",
+    "web-vitals": "^2.1.4",
+    "workbox-background-sync": "^6.5.4",
+    "workbox-broadcast-update": "^6.5.4",
+    "workbox-cacheable-response": "^6.5.4",
+    "workbox-core": "^6.5.4",
+    "workbox-expiration": "^6.5.4",
+    "workbox-google-analytics": "^6.5.4",
+    "workbox-navigation-preload": "^6.5.4",
+    "workbox-precaching": "^6.5.4",
+    "workbox-range-requests": "^6.5.4",
+    "workbox-routing": "^6.5.4",
+    "workbox-strategies": "^6.5.4",
+    "workbox-streams": "^6.5.4"
   },
   "devDependencies": {
-    "ajv": "^6.9.1",
+    "ajv": "^8.11.0",
     "cross-env": "^7.0.3",
-    "eslint": "^7.25.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.7.2",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.23.2",
-    "nan": "^2.14.2",
-    "typescript": "^4.2.4"
+    "eslint": "^8.22.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-react": "^7.30.1",
+    "nan": "^2.16.0",
+    "typescript": "^4.7.4"
+  },
+  "overrides": {
+    "autoprefixer": "10.4.5"
+  },
+  "resolutions": {
+    "css-what": "^5.0.1",
+    "nth-check": "^3.0.1"
   },
   "scripts": {
-"prestart": "node aspnetcore-https && node aspnetcore-react",
+    "prestart": "node aspnetcore-https && node aspnetcore-react",
     "start": "rimraf ./build && react-scripts start",
     "build": "react-scripts build",
     "test": "cross-env CI=true react-scripts test --env=jsdom",
