@@ -1,13 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { MemoryRouter } from 'react-router-dom';
 import App from './App';
 
 it('renders without crashing', async () => {
   const div = document.createElement('div');
-  ReactDOM.render(
+  const root = createRoot(div);
+  root.render(
     <MemoryRouter>
       <App />
-    </MemoryRouter>, div);
+    </MemoryRouter>);
   await new Promise(resolve => setTimeout(resolve, 1000));
 });
