@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 export class Home extends Component {
   static displayName = Home.name;
 
-  render () {
+  render() {
     return (
       <div>
         <h1>Hello, world!</h1>
