@@ -5,11 +5,11 @@ import { NavMenu } from './NavMenu';
 export class Layout extends Component {
   static displayName = Layout.name;
 
-  render () {
+  render() {
     return (
       <div>
         <NavMenu />
-        <Container>
+        <Container tag="main">
           {this.props.children}
         </Container>
       </div>
