@@ -1,17 +1,29 @@
-const createProxyMiddleware = require('http-proxy-middleware');
+const { createProxyMiddleware } = require('http-proxy-middleware');
 const { env } = require('process');
 
 const target = env.ASPNETCORE_HTTPS_PORT ? `https://localhost:${env.ASPNETCORE_HTTPS_PORT}` :
-  env.ASPNETCORE_URLS ? env.ASPNETCORE_URLS.split(';')[0] : 'http://localhost:35790';
+  env.ASPNETCORE_URLS ? env.ASPNETCORE_URLS.split(';')[0] : 'http://localhost:15140';
 
-const context =  [
+const context = [
   "/weatherforecast",
 ];
 
-module.exports = function(app) {
+const onError = (err, req, resp, target) => {
+    console.error(`${err.message}`);
+}
+
+module.exports = function (app) {
   const appProxy = createProxyMiddleware(context, {
     target: target,
-    secure: false
+    // Handle errors to prevent the proxy middleware from crashing when
+    // the ASP NET Core webserver is unavailable
+    onError: onError,
+    secure: false,
+    // Uncomment this line to add support for proxying websockets
+    //ws: true, 
+    headers: {
+      Connection: 'Keep-Alive'
+    }
   });
 
   app.use(appProxy);
